(Many changes)
State methods are now executed in the context of the state in which they are declared, rather than the owner, which must now be accessed as ; most notably this provides state methods with a more robust inheritance model, e.g. using . However, owner methods, i.e. methods that were originally direct properties of the owner object, are still executed as usual, in the context of the owner.

State.Controller.js
- Moved function body of  to , to reduce memory footprint; similarly with  and
- Renamed  to , added
- Renamed  closure var from  to
State.Transition.js
- Changed  to , added optional  param

State.js
- Renamed  closure var from  to
- Added  and rewrote  and  to differentiate state vs. owner methods as noted above

pre.js
- Added

Misc
- Numerous comment blocks added