add equality check for `state` in accessor
This allows constructors with a stateful prototype to cater to engines
that optimize for shape order, by including something akin to:

    this.state = this.constructor.prototype.state;

On first invocation the inherited `state` method will be replaced with
the object’s own accessor method, as usual, but now the object’s
shape/hidden class/etc. will be deterministic with respect to the
`state` property.